NEW - bug 235754: task list externalizer error in log on startup https://bugs.eclipse.org/bugs/show_bug.cgi?id=235754

+++ org.eclipse.mylyn.tasks.core/src/org/eclipse/mylyn/internal/tasks/core/externalization/AbstractExternalizationParticipant.java
@@ -55,7 +55,10 @@ public abstract class AbstractExternalizationParticipant implements IExternaliza
 					+ originalFile.getName());
 			originalFile.renameTo(failed);
 		}
-		return backup.renameTo(originalFile);
+		if (backup.exists()) {
+			return backup.renameTo(originalFile);
+		}
+		return false;
 	}
 
 	protected boolean takeSnapshot(File file) {
@@ -84,9 +87,10 @@ public abstract class AbstractExternalizationParticipant implements IExternaliza
 				load(context.getRootPath(), monitor);
 			} catch (CoreException e) {
 				if (dataFile != null) {
-					restoreSnapshot(dataFile);
+					if (restoreSnapshot(dataFile)) {
+						load(context.getRootPath(), monitor);
+					}
 				}
-				load(context.getRootPath(), monitor);
 			}
 			break;
 		case SNAPSHOT:
